code_samples/multi_agent/02_orchestration/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0
pydantic>=2.0.0
pytest>=7.0.0